Remove tailwind build artifacts
