--- postcss.config.cjs
@@ -1,11 +1,0 @@
-const tailwindcss = require("tailwindcss");
-const autoprefixer = require("autoprefixer");
-
-const config = {
-  plugins: [
-    autoprefixer,
-    tailwindcss(),
-  ],
-};
-
-module.exports = config;
