+++ svelte.config.js
@@ -3,12 +3,7 @@ import preprocess from "svelte-preprocess";
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-  preprocess: [
-    preprocess({
-      postcss: true,
-    }),
-  ],
-
+  preprocess: preprocess(),
   kit: {
     adapter: adapter(),
     package: {
